new maj libft

+++ ft_memcpy.c
@@ -6,7 +6,7 @@
 /*   By: qstemper <qstemper@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2015/11/09 10:47:12 by qstemper          #+#    #+#             */
-/*   Updated: 2015/11/09 11:03:33 by qstemper         ###   ########.fr       */
+/*   Updated: 2015/11/24 12:22:34 by qstemper         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,10 +14,10 @@
 
 void	*ft_memcpy(void *dest, const void *src, size_t n)
 {
-	int		i;
+	size_t	i;
 
 	i = 0;
-	while(i < n)
+	while (i < n)
 	{
 		((unsigned char *)dest)[i] = ((unsigned char *)src)[i];
 		i++;
